使用id为唯一标识改为使用unionId，integer改为long

+++ socialuni-server/socialuni-admin-sdk/src/main/java/com/socialuni/social/admin/controller/ReportController.java
@@ -80,7 +80,7 @@ public class ReportController {
     }
 
     @PostMapping("queryUserReports")
-    public ResultRO<List<ReportRO>> queryReports(Integer userId) {
+    public ResultRO<List<ReportRO>> queryReports(Long userId) {
         //展示用户的前多少条动态，评论，
 
         /*SocialuniUserDo user = new SocialuniUserDo();
